backend/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2


Need to uninstall the bellow packages.
IPython-7.24.1 PyWavelets-1.1.1 absl-py-0.13.0 astunparse-1.6.3 backcall-0.2.0 beautifulsoup4-4.9.3 cachetools-4.2.2 certifi-2021.5.30 cffi-1.14.5 chardet-4.0.0 cryptography-3.4.7 cycler-0.10.0 decorator-5.0.9 distro-1.5.0 flatbuffers-1.12 gast-0.4.0 google-auth-1.31.0 google-auth-oauthlib-0.4.4 google-pasta-0.2.0 grpcio-1.34.1 h5py-3.2.1 idna-2.10 imageio-2.9.0 ipython-genutils-0.2.0 jedi-0.18.0 joblib-1.0.1 keras-2.4.3 keras-nightly-2.5.0.dev2021032900 keras-preprocessing-1.1.2 kiwisolver-1.3.1 markdown-3.3.4 matplotlib-3.4.2 matplotlib-inline-0.1.2 networkx-2.5.1 numpy-1.20.3 oauthlib-3.1.1 opt-einsum-3.3.0 pandas-1.2.4 parso-0.8.2 pdfminer.six-20201018 pdftotree-0.5.0 pexpect-4.8.0 pickleshare-0.7.5 prompt-toolkit-3.0.18 protobuf-3.17.3 ptyprocess-0.7.0 pyasn1-0.4.8 pyasn1-modules-0.2.8 pycparser-2.20 pygments-2.9.0 pyparsing-2.4.7 python-dateutil-2.8.1 pyyaml-5.4.1 requests-2.25.1 requests-oauthlib-1.3.0 rsa-4.7.2 scikit-image-0.18.1 scikit-learn-0.24.2 scipy-1.6.3 selectivesearch-0.4 six-1.15.0 sklearn-0.0 sortedcontainers-2.4.0 soupsieve-2.2.1 tabula-py-2.2.0 tensorboard-2.5.0 tensorboard-data-server-0.6.1 tensorboard-plugin-wit-1.8.0 tensorflow-2.5.0 tensorflow-estimator-2.5.0 termcolor-1.1.0 threadpoolctl-2.1.0 tifffile-2021.6.14 traitlets-5.0.5 typing-extensions-3.7.4.3 urllib3-1.26.5 wand-0.6.6 wcwidth-0.2.5 werkzeug-2.0.1 wheel-0.36.2 wrapt-1.12.1